+++ db/migrate/20090627061643_create_funds.rb
@@ -2,7 +2,7 @@ class CreateFunds < ActiveRecord::Migration
   def self.up
     create_table :funds do |t|
       t.string :name
-      t.integer :initial_balance_in_cents, :default => 0
+      t.float :initial_balance_in_cents, :default => 0
       t.boolean :default_synchronize_fund, :default => false
       t.integer :user_id
